+++ src/Digdir.Domain.Dialogporten.Application/Common/Behaviours/DataLoader/IRequireDataLoader.cs
@@ -1,0 +1,6 @@
+namespace Digdir.Domain.Dialogporten.Application.Common.Behaviours.DataLoader;
+
+internal interface IRequireDataLoader
+{
+    Dictionary<string, object?> PreloadedData { get; }
+}
